refactor: remove former NB member scraper; consolidate PE functions into provincial/pe.go
Agent-Logs-Url: https://github.com/philspins/opendocket/sessions/0206cb27-bf37-4dac-a881-6aea030b39bc
