+++ internal/scraper/members.go
@@ -583,73 +583,6 @@ func CrawlNewBrunswickMembersFromWebsite(indexURL string, client *http.Client) (
 	return profiles, nil
 }
 
-// NBFormerMembersURL is the canonical URL for former NB MLAs.
-const NBFormerMembersURL = "https://www.legnb.ca/en/members/former"
-
-// CrawlNewBrunswickFormerMembersFromWebsite scrapes former MLA profiles from the
-// New Brunswick Legislative Assembly website. Former members appear in votes from
-// previous sessions but are not present in the Represent API.
-func CrawlNewBrunswickFormerMembersFromWebsite(indexURL string, client *http.Client) ([]MemberProfile, error) {
-	if indexURL == "" {
-		indexURL = NBFormerMembersURL
-	}
-	if client == nil {
-		client = utils.NewHTTPClient()
-	}
-	doc, err := fetchDoc(indexURL, client)
-	if err != nil {
-		return nil, fmt.Errorf("nb former members website: %w", err)
-	}
-
-	var profiles []MemberProfile
-	doc.Find("div.member-card").Each(func(_ int, card *goquery.Selection) {
-		nameEl := card.Find("li.member-card-description-name a[href]")
-		href, _ := nameEl.Attr("href")
-		rawName := strings.TrimSpace(nameEl.Find("h3").Text())
-		if rawName == "" {
-			return
-		}
-
-		name := nbConvertMemberName(rawName)
-
-		nameSlug := urlLastSegment(href)
-		if nameSlug == "" {
-			nameSlug = nameToSlug(name)
-		}
-		memberID := "nb-legislature-" + nameSlug
-
-		partyLi := card.Find("li.member-card-description-party")
-		partyLi.Find("div").Remove()
-		party := strings.TrimSpace(partyLi.Text())
-
-		riding := strings.TrimSpace(card.Find("li.member-card-description-riding span").Text())
-
-		photoSrc := strings.TrimSpace(card.Find("div.member-card-avatar img").AttrOr("src", ""))
-		photoSrc = strings.ReplaceAll(photoSrc, "\\", "/")
-		var photoURL string
-		if photoSrc != "" {
-			photoURL = resolveRelativeURL(indexURL, photoSrc)
-		}
-
-		profiles = append(profiles, MemberProfile{
-			ID:              memberID,
-			Name:            name,
-			Party:           party,
-			Riding:          riding,
-			Province:        "New Brunswick",
-			Role:            "MLA",
-			PhotoURL:        photoURL,
-			Chamber:         "legislature",
-			Active:          false,
-			LastScraped:     utils.NowISO(),
-			GovernmentLevel: "provincial",
-		})
-	})
-
-	clog.Infof("[members] fetched %d former NB members from website", len(profiles))
-	return profiles, nil
-}
-
 // nbConvertMemberName converts a name in "LastName, FirstName" format to
 // "FirstName LastName" for consistent storage.  It is used for both New
 // Brunswick (NB) member pages and Newfoundland and Labrador (NL) member-index
